Major Update 3, Update form pendaftaran, admin connect database, authentication, hashing

+++ app/api/admins/[id]/route.ts
@@ -1,17 +1,29 @@
 import { NextResponse } from "next/server";
 import prisma from "@/lib/prisma";
 
-export async function DELETE(request: Request, { params }: { params: { id: string } }) {
+export async function DELETE(
+  request: Request,
+  { params }: { params: Promise<{ id: string }> } // <-- Ubah Tipe Jadi Promise
+) {
   try {
-    const id = params.id;
+    const { id } = await params; // <-- WAJIB DI-AWAIT DI NEXT.JS 15
 
-    await prisma.admin.delete({
-      where: { id }
+    // Cek admin
+    const admin = await prisma.admin.findUnique({
+      where: { id: id }, 
     });
 
-    return NextResponse.json({ message: "Berhasil dihapus" });
+    if (!admin) {
+      return NextResponse.json({ error: "Admin tidak ditemukan" }, { status: 404 });
+    }
 
+    await prisma.admin.delete({
+      where: { id: id },
+    });
+
+    return NextResponse.json({ message: "Admin berhasil dihapus" });
   } catch (error) {
+    console.error("Delete Admin Error:", error);
     return NextResponse.json({ error: "Gagal menghapus admin" }, { status: 500 });
   }
 }
